% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggldplot.R
\name{make_full_dl_matrix}
\alias{make_full_dl_matrix}
\title{Make full partial martix. For internal use.}
\usage{
make_full_dl_matrix(x)
}
\arguments{
\item{x}{A vcftools ld data.frame}
}
\value{
a data.frame
}
\description{
Make full partial martix. For internal use.
}

